// src/common/pagination/interfaces/pagination-result.interface.ts
export interface PaginationResult<T> {
    data: T[];
    meta: {
        itemsPerPage: number;
        totalItems: number;
        currentPage: number;
        totalPages: number;
    };
    links: {
        first: string;
        last:string;
        current: string;
        next: string;
        previous: string;
    }
}